Added special class DYNABLOB that stands for Dynamic Binary Large OBject. This type uses dynamic memory to store variable data types into a void pointer that can later be reinterpreted as whatever you want. We can track the type of data stored using the TYPE within, immutable and query only. Its an enumeration we can work off of to decide what the last interpretation of the data was.
Using a reinterpretation cast like so before the .getRaw(), like so:
*(int*) dynablob.getRaw()

We can interpret the data within as an integer easily. this is fully conversion proven, where converting from TYPE > dynablob > TYPE should always work correctly.
Technically any data can be stored in here, as long as malloc() can allocate memory for it.